+++ bindings/dragonruby/cexts/ice_cpu_drb.c
@@ -8,49 +8,20 @@
 #define ICE_CPU_IMPL 1
 #include "ice_cpu.h"
 
-// MRuby `typedef`s mrb_int in the mruby/value.h
-// Then `#define`s mrb_int in mruby.h
-// We need to undo the macro and avoid it's usage
-// FIXME: I'm surely doing something wrong
-#ifdef mrb_int
-#undef mrb_int
-#endif
+static drb_api_t *drb_api;
 
-void *(*drb_symbol_lookup)(const char *sym) = NULL;
-
-static void (*drb_free_foreign_object_f)(mrb_state *, void *);
-static void (*drb_typecheck_float_f)(mrb_state *, mrb_value);
-static void (*drb_typecheck_int_f)(mrb_state *, mrb_value);
-static void (*drb_typecheck_bool_f)(mrb_state *, mrb_value);
-static void (*drb_typecheck_aggregate_f)(mrb_state *, mrb_value, struct RClass *, mrb_data_type *);
-static struct RClass *(*mrb_module_get_f)(mrb_state *, const char *);
-static mrb_int (*mrb_get_args_f)(mrb_state *, mrb_args_format, ...);
-static struct RClass *(*mrb_module_get_under_f)(mrb_state *, struct RClass *, const char *);
-static struct RClass *(*mrb_class_get_under_f)(mrb_state *, struct RClass *, const char *);
-static struct RClass *(*mrb_define_module_under_f)(mrb_state *, struct RClass *, const char *);
-static void (*mrb_define_module_function_f)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec);
-static struct RClass *(*mrb_define_class_under_f)(mrb_state *, struct RClass *, const char *, struct RClass *);
-static void (*mrb_define_method_f)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec);
-static void (*mrb_define_class_method_f)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec);
-static struct RData *(*mrb_data_object_alloc_f)(mrb_state *, struct RClass *, void *, const mrb_data_type *);
-static mrb_value (*mrb_str_new_cstr_f)(mrb_state *, const char *);
-static void (*mrb_raise_f)(mrb_state *, struct RClass *, const char *);
-static void (*mrb_raisef_f)(mrb_state *, struct RClass *, const char *, ...);
-static mrb_value (*drb_float_value_f)(mrb_state *, mrb_float);
-static struct RClass *(*drb_getruntime_error_f)(mrb_state *);
-static struct RClass *(*drb_getargument_error_f)(mrb_state *);
 static void drb_free_foreign_object_indirect(mrb_state *state, void *pointer) {
-    drb_free_foreign_object_f(state, pointer);
+    drb_api->drb_free_foreign_object(state, pointer);
 }
 static int drb_ffi__ZTSi_FromRuby(mrb_state *state, mrb_value self) {
-    drb_typecheck_int_f(state, self);
+    drb_api->drb_typecheck_int(state, self);
     return mrb_fixnum(self);
 }
 static mrb_value drb_ffi__ZTSi_ToRuby(mrb_state *state, int value) {
     return mrb_fixnum_value(value);
 }
 static ice_cpu_bool drb_ffi__ZTS12ice_cpu_bool_FromRuby(mrb_state *state, mrb_value self) {
-    drb_typecheck_int_f(state, self);
+    drb_api->drb_typecheck_int(state, self);
     return mrb_fixnum(self);
 }
 static mrb_value drb_ffi__ZTS12ice_cpu_bool_ToRuby(mrb_state *state, ice_cpu_bool value) {
@@ -65,42 +36,42 @@ static mrb_data_type ForeignObjectType_ZTSP12ice_cpu_info = {"ice_cpu_info*", dr
 static ice_cpu_info *drb_ffi__ZTSP12ice_cpu_info_FromRuby(mrb_state *state, mrb_value self) {
     if (mrb_nil_p(self))
         return 0;
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "ICE_CPU_INFO_PTR");
-    drb_typecheck_aggregate_f(state, self, klass, &ForeignObjectType_ZTSP12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "ICE_CPU_INFO_PTR");
+    drb_api->drb_typecheck_aggregate(state, self, klass, &ForeignObjectType_ZTSP12ice_cpu_info);
     return ((struct drb_foreign_object_ZTSP12ice_cpu_info *)DATA_PTR(self))->value;
 }
 static mrb_value drb_ffi__ZTSP12ice_cpu_info_ToRuby(mrb_state *state, ice_cpu_info *value) {
     struct drb_foreign_object_ZTSP12ice_cpu_info *ptr = calloc(1, sizeof(struct drb_foreign_object_ZTSP12ice_cpu_info));
     ptr->value = value;
     ptr->kind = drb_foreign_object_kind_pointer;
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "ICE_CPU_INFO_PTR");
-    struct RData *rdata = mrb_data_object_alloc_f(state, klass, ptr, &ForeignObjectType_ZTSP12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "ICE_CPU_INFO_PTR");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(state, klass, ptr, &ForeignObjectType_ZTSP12ice_cpu_info);
     return mrb_obj_value(rdata);
 }
 struct drb_foreign_object_ZTS12ice_cpu_info {
     drb_foreign_object_kind kind;
     ice_cpu_info value;
 };
-static mrb_data_type ForeignObjectType_ZTS12ice_cpu_info = {"ICE_CPU_INFO", drb_free_foreign_object_indirect};
+static mrb_data_type ForeignObjectType_ZTS12ice_cpu_info = {"ice_cpu_info", drb_free_foreign_object_indirect};
 static ice_cpu_info drb_ffi__ZTS12ice_cpu_info_FromRuby(mrb_state *state, mrb_value self) {
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "ICE_CPU_INFO");
-    drb_typecheck_aggregate_f(state, self, klass, &ForeignObjectType_ZTS12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "ICE_CPU_INFO");
+    drb_api->drb_typecheck_aggregate(state, self, klass, &ForeignObjectType_ZTS12ice_cpu_info);
     return ((struct drb_foreign_object_ZTS12ice_cpu_info *)DATA_PTR(self))->value;
 }
 static mrb_value drb_ffi__ZTS12ice_cpu_info_ToRuby(mrb_state *state, ice_cpu_info value) {
     struct drb_foreign_object_ZTS12ice_cpu_info *ptr = calloc(1, sizeof(struct drb_foreign_object_ZTS12ice_cpu_info));
     ptr->value = value;
     ptr->kind = drb_foreign_object_kind_struct;
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "ICE_CPU_INFO");
-    struct RData *rdata = mrb_data_object_alloc_f(state, klass, ptr, &ForeignObjectType_ZTS12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "ICE_CPU_INFO");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(state, klass, ptr, &ForeignObjectType_ZTS12ice_cpu_info);
     return mrb_obj_value(rdata);
 }
 struct drb_foreign_object_ZTSPc {
@@ -114,31 +85,31 @@ static char *drb_ffi__ZTSPc_FromRuby(mrb_state *state, mrb_value self) {
         return 0;
     if (mrb_type(self) == MRB_TT_STRING)
         return RSTRING_PTR(self);
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "CharPointer");
-    drb_typecheck_aggregate_f(state, self, klass, &ForeignObjectType_ZTSPc);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "CharPointer");
+    drb_api->drb_typecheck_aggregate(state, self, klass, &ForeignObjectType_ZTSPc);
     return ((struct drb_foreign_object_ZTSPc *)DATA_PTR(self))->value;
 }
 static mrb_value drb_ffi__ZTSPc_ToRuby(mrb_state *state, char *value) {
     struct drb_foreign_object_ZTSPc *ptr = calloc(1, sizeof(struct drb_foreign_object_ZTSPc));
     ptr->value = value;
     ptr->kind = drb_foreign_object_kind_pointer;
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "CharPointer");
-    struct RData *rdata = mrb_data_object_alloc_f(state, klass, ptr, &ForeignObjectType_ZTSPc);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "CharPointer");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(state, klass, ptr, &ForeignObjectType_ZTSPc);
     return mrb_obj_value(rdata);
 }
 static unsigned int drb_ffi__ZTSj_FromRuby(mrb_state *state, mrb_value self) {
-    drb_typecheck_int_f(state, self);
+    drb_api->drb_typecheck_int(state, self);
     return mrb_fixnum(self);
 }
 static mrb_value drb_ffi__ZTSj_ToRuby(mrb_state *state, unsigned int value) {
     return mrb_fixnum_value(value);
 }
 static char drb_ffi__ZTSc_FromRuby(mrb_state *state, mrb_value self) {
-    drb_typecheck_int_f(state, self);
+    drb_api->drb_typecheck_int(state, self);
     return mrb_fixnum(self);
 }
 static mrb_value drb_ffi__ZTSc_ToRuby(mrb_state *state, char value) {
@@ -149,10 +120,10 @@ static mrb_value drb_ffi__ZTSP12ice_cpu_info_New(mrb_state *mrb, mrb_value self)
     ptr->kind = drb_foreign_object_kind_pointer;
     ptr->value = calloc(1, sizeof(ice_cpu_info));
     ptr->should_free = 1;
-    struct RClass *FFI = mrb_module_get_f(mrb, "FFI");
-    struct RClass *module = mrb_module_get_under_f(mrb, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(mrb, module, "ICE_CPU_INFO_PTR");
-    struct RData *rdata = mrb_data_object_alloc_f(mrb, klass, ptr, &ForeignObjectType_ZTSP12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(mrb, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(mrb, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(mrb, module, "ICE_CPU_INFO_PTR");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(mrb, klass, ptr, &ForeignObjectType_ZTSP12ice_cpu_info);
     return mrb_obj_value(rdata);
 }
 static mrb_value drb_ffi__ZTSP12ice_cpu_info_GetValue(mrb_state *mrb, mrb_value value) {
@@ -167,14 +138,14 @@ static mrb_value drb_ffi__ZTSP12ice_cpu_info_IsNil(mrb_state *state, mrb_value s
 static mrb_value drb_ffi__ZTSP12ice_cpu_info_GetAt(mrb_state *mrb, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(mrb, "*", &args, &argc);
+    drb_api->mrb_get_args(mrb, "*", &args, &argc);
     int index = drb_ffi__ZTSi_FromRuby(mrb, args[0]);
     return drb_ffi__ZTS12ice_cpu_info_ToRuby(mrb, drb_ffi__ZTSP12ice_cpu_info_FromRuby(mrb, self)[index]);
 }
 static mrb_value drb_ffi__ZTSP12ice_cpu_info_SetAt(mrb_state *mrb, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(mrb, "*", &args, &argc);
+    drb_api->mrb_get_args(mrb, "*", &args, &argc);
     int index = drb_ffi__ZTSi_FromRuby(mrb, args[0]);
     ice_cpu_info new_value = drb_ffi__ZTS12ice_cpu_info_FromRuby(mrb, args[1]);
     drb_ffi__ZTSP12ice_cpu_info_FromRuby(mrb, self)[index] = new_value;
@@ -185,10 +156,10 @@ static mrb_value drb_ffi__ZTSPc_New(mrb_state *mrb, mrb_value self) {
     ptr->kind = drb_foreign_object_kind_pointer;
     ptr->value = calloc(1, sizeof(char));
     ptr->should_free = 1;
-    struct RClass *FFI = mrb_module_get_f(mrb, "FFI");
-    struct RClass *module = mrb_module_get_under_f(mrb, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(mrb, module, "CharPointer");
-    struct RData *rdata = mrb_data_object_alloc_f(mrb, klass, ptr, &ForeignObjectType_ZTSPc);
+    struct RClass *FFI = drb_api->mrb_module_get(mrb, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(mrb, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(mrb, module, "CharPointer");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(mrb, klass, ptr, &ForeignObjectType_ZTSPc);
     return mrb_obj_value(rdata);
 }
 static mrb_value drb_ffi__ZTSPc_GetValue(mrb_state *mrb, mrb_value value) {
@@ -203,28 +174,28 @@ static mrb_value drb_ffi__ZTSPc_IsNil(mrb_state *state, mrb_value self) {
 static mrb_value drb_ffi__ZTSPc_GetAt(mrb_state *mrb, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(mrb, "*", &args, &argc);
+    drb_api->mrb_get_args(mrb, "*", &args, &argc);
     int index = drb_ffi__ZTSi_FromRuby(mrb, args[0]);
     return drb_ffi__ZTSc_ToRuby(mrb, drb_ffi__ZTSPc_FromRuby(mrb, self)[index]);
 }
 static mrb_value drb_ffi__ZTSPc_SetAt(mrb_state *mrb, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(mrb, "*", &args, &argc);
+    drb_api->mrb_get_args(mrb, "*", &args, &argc);
     int index = drb_ffi__ZTSi_FromRuby(mrb, args[0]);
     char new_value = drb_ffi__ZTSc_FromRuby(mrb, args[1]);
     drb_ffi__ZTSPc_FromRuby(mrb, self)[index] = new_value;
     return mrb_nil_value();
 }
 static mrb_value drb_ffi__ZTSPc_GetString(mrb_state *state, mrb_value self) {
-    return mrb_str_new_cstr_f(state, drb_ffi__ZTSPc_FromRuby(state, self));
+    return drb_api->mrb_str_new_cstr(state, drb_ffi__ZTSPc_FromRuby(state, self));
 }
 static mrb_value drb_ffi__ZTS12ice_cpu_info_New(mrb_state *state, mrb_value self) {
     struct drb_foreign_object_ZTS12ice_cpu_info *ptr = calloc(1, sizeof(struct drb_foreign_object_ZTS12ice_cpu_info *));
-    struct RClass *FFI = mrb_module_get_f(state, "FFI");
-    struct RClass *module = mrb_module_get_under_f(state, FFI, "CExt");
-    struct RClass *klass = mrb_class_get_under_f(state, module, "ICE_CPU_INFO");
-    struct RData *rdata = mrb_data_object_alloc_f(state, klass, ptr, &ForeignObjectType_ZTS12ice_cpu_info);
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_module_get_under(state, FFI, "CExt");
+    struct RClass *klass = drb_api->mrb_class_get_under(state, module, "ICE_CPU_INFO");
+    struct RData *rdata = drb_api->mrb_data_object_alloc(state, klass, ptr, &ForeignObjectType_ZTS12ice_cpu_info);
     return mrb_obj_value(rdata);
 }
 static mrb_value drb_ffi__ZTS12ice_cpu_info_name_Get(mrb_state *state, mrb_value self) {
@@ -234,7 +205,7 @@ static mrb_value drb_ffi__ZTS12ice_cpu_info_name_Get(mrb_state *state, mrb_value
 static mrb_value drb_ffi__ZTS12ice_cpu_info_name_Set(mrb_state *state, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(state, "*", &args, &argc);
+    drb_api->mrb_get_args(state, "*", &args, &argc);
     char *new_value = drb_ffi__ZTSPc_FromRuby(state, args[0]);
     (&((struct drb_foreign_object_ZTS12ice_cpu_info *)DATA_PTR(self))->value)->name = new_value;
     return mrb_nil_value();
@@ -246,7 +217,7 @@ static mrb_value drb_ffi__ZTS12ice_cpu_info_cores_Get(mrb_state *state, mrb_valu
 static mrb_value drb_ffi__ZTS12ice_cpu_info_cores_Set(mrb_state *state, mrb_value self) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(state, "*", &args, &argc);
+    drb_api->mrb_get_args(state, "*", &args, &argc);
     unsigned int new_value = drb_ffi__ZTSj_FromRuby(state, args[0]);
     (&((struct drb_foreign_object_ZTS12ice_cpu_info *)DATA_PTR(self))->value)->cores = new_value;
     return mrb_nil_value();
@@ -254,63 +225,37 @@ static mrb_value drb_ffi__ZTS12ice_cpu_info_cores_Set(mrb_state *state, mrb_valu
 static mrb_value drb_ffi_ice_cpu_get_info_Binding(mrb_state *state, mrb_value value) {
     mrb_value *args = 0;
     mrb_int argc = 0;
-    mrb_get_args_f(state, "*", &args, &argc);
+    drb_api->mrb_get_args(state, "*", &args, &argc);
     if (argc != 1)
-        mrb_raisef_f(state, drb_getargument_error_f(state), "'ice_cpu_get_info': wrong number of arguments (%d for 1)", argc);
+        drb_api->mrb_raisef(state, drb_api->drb_getargument_error(state), "'ice_cpu_get_info': wrong number of arguments (%d for 1)", argc);
     ice_cpu_info *cpu_info_0 = drb_ffi__ZTSP12ice_cpu_info_FromRuby(state, args[0]);
     ice_cpu_bool ret_val = ice_cpu_get_info(cpu_info_0);
     return drb_ffi__ZTS12ice_cpu_bool_ToRuby(state, ret_val);
 }
-static int drb_ffi_init_indirect_functions(void *(*lookup)(const char *));
 DRB_FFI_EXPORT
-void drb_register_c_extensions(void *(*lookup)(const char *), mrb_state *state, struct RClass *FFI) {
-    if (drb_ffi_init_indirect_functions(lookup))
-        return;
-    struct RClass *module = mrb_define_module_under_f(state, FFI, "CExt");
+void drb_register_c_extensions_with_api(mrb_state *state, struct drb_api_t *api) {
+    drb_api = api;
+    struct RClass *FFI = drb_api->mrb_module_get(state, "FFI");
+    struct RClass *module = drb_api->mrb_define_module_under(state, FFI, "CExt");
     struct RClass *object_class = state->object_class;
-    mrb_define_module_function_f(state, module, "ice_cpu_get_info", drb_ffi_ice_cpu_get_info_Binding, MRB_ARGS_REQ(1));
-    struct RClass *ice_cpu_info_ptr_class = mrb_define_class_under_f(state, module, "ICE_CPU_INFO_PTR", object_class);
-    mrb_define_class_method_f(state, ice_cpu_info_ptr_class, "new", drb_ffi__ZTSP12ice_cpu_info_New, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, ice_cpu_info_ptr_class, "value", drb_ffi__ZTSP12ice_cpu_info_GetValue, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, ice_cpu_info_ptr_class, "[]", drb_ffi__ZTSP12ice_cpu_info_GetAt, MRB_ARGS_REQ(1));
-    mrb_define_method_f(state, ice_cpu_info_ptr_class, "[]=", drb_ffi__ZTSP12ice_cpu_info_SetAt, MRB_ARGS_REQ(2));
-    mrb_define_method_f(state, ice_cpu_info_ptr_class, "nil?", drb_ffi__ZTSP12ice_cpu_info_IsNil, MRB_ARGS_REQ(0));
-    struct RClass *CharPointerClass = mrb_define_class_under_f(state, module, "CharPointer", object_class);
-    mrb_define_class_method_f(state, CharPointerClass, "new", drb_ffi__ZTSPc_New, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, CharPointerClass, "value", drb_ffi__ZTSPc_GetValue, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, CharPointerClass, "[]", drb_ffi__ZTSPc_GetAt, MRB_ARGS_REQ(1));
-    mrb_define_method_f(state, CharPointerClass, "[]=", drb_ffi__ZTSPc_SetAt, MRB_ARGS_REQ(2));
-    mrb_define_method_f(state, CharPointerClass, "nil?", drb_ffi__ZTSPc_IsNil, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, CharPointerClass, "str", drb_ffi__ZTSPc_GetString, MRB_ARGS_REQ(0));
-    struct RClass *ice_cpu_info_class = mrb_define_class_under_f(state, module, "ICE_CPU_INFO", object_class);
-    mrb_define_class_method_f(state, ice_cpu_info_class, "new", drb_ffi__ZTS12ice_cpu_info_New, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, ice_cpu_info_class, "name", drb_ffi__ZTS12ice_cpu_info_name_Get, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, ice_cpu_info_class, "name=", drb_ffi__ZTS12ice_cpu_info_name_Set, MRB_ARGS_REQ(1));
-    mrb_define_method_f(state, ice_cpu_info_class, "cores", drb_ffi__ZTS12ice_cpu_info_cores_Get, MRB_ARGS_REQ(0));
-    mrb_define_method_f(state, ice_cpu_info_class, "cores=", drb_ffi__ZTS12ice_cpu_info_cores_Set, MRB_ARGS_REQ(1));
-}
-static int drb_ffi_init_indirect_functions(void *(*lookup)(const char *fnname)) {
-  drb_symbol_lookup = lookup;
-  if (!(drb_float_value_f = (mrb_value (*)(mrb_state *, mrb_float)) lookup("drb_float_value"))) return -1;
-  if (!(drb_free_foreign_object_f = (void (*)(mrb_state *, void *)) lookup("drb_free_foreign_object"))) return -1;
-  if (!(drb_getargument_error_f = (struct RClass *(*)(mrb_state *)) lookup("drb_getargument_error"))) return -1;
-  if (!(drb_getruntime_error_f = (struct RClass *(*)(mrb_state *)) lookup("drb_getruntime_error"))) return -1;
-  if (!(drb_typecheck_aggregate_f = (void (*)(mrb_state *, mrb_value, struct RClass *, mrb_data_type *)) lookup("drb_typecheck_aggregate"))) return -1;
-  if (!(drb_typecheck_bool_f = (void (*)(mrb_state *, mrb_value)) lookup("drb_typecheck_bool"))) return -1;
-  if (!(drb_typecheck_float_f = (void (*)(mrb_state *, mrb_value)) lookup("drb_typecheck_float"))) return -1;
-  if (!(drb_typecheck_int_f = (void (*)(mrb_state *, mrb_value)) lookup("drb_typecheck_int"))) return -1;
-  if (!(mrb_class_get_under_f = (struct RClass *(*)(mrb_state *, struct RClass *, const char *)) lookup("mrb_class_get_under"))) return -1;
-  if (!(mrb_data_object_alloc_f = (struct RData *(*)(mrb_state *, struct RClass *, void *, const mrb_data_type *)) lookup("mrb_data_object_alloc"))) return -1;
-  if (!(mrb_define_class_method_f = (void (*)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec)) lookup("mrb_define_class_method"))) return -1;
-  if (!(mrb_define_class_under_f = (struct RClass *(*)(mrb_state *, struct RClass *, const char *, struct RClass *)) lookup("mrb_define_class_under"))) return -1;
-  if (!(mrb_define_method_f = (void (*)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec)) lookup("mrb_define_method"))) return -1;
-  if (!(mrb_define_module_function_f = (void (*)(mrb_state *, struct RClass *, const char *, mrb_func_t, mrb_aspec)) lookup("mrb_define_module_function"))) return -1;
-  if (!(mrb_define_module_under_f = (struct RClass *(*)(mrb_state *, struct RClass *, const char *)) lookup("mrb_define_module_under"))) return -1;
-  if (!(mrb_get_args_f = (mrb_int (*)(mrb_state *, mrb_args_format, ...)) lookup("mrb_get_args"))) return -1;
-  if (!(mrb_module_get_f = (struct RClass *(*)(mrb_state *, const char *)) lookup("mrb_module_get"))) return -1;
-  if (!(mrb_module_get_under_f = (struct RClass *(*)(mrb_state *, struct RClass *, const char *)) lookup("mrb_module_get_under"))) return -1;
-  if (!(mrb_raise_f = (void (*)(mrb_state *, struct RClass *, const char *)) lookup("mrb_raise"))) return -1;
-  if (!(mrb_raisef_f = (void (*)(mrb_state *, struct RClass *, const char *, ...)) lookup("mrb_raisef"))) return -1;
-  if (!(mrb_str_new_cstr_f = (mrb_value (*)(mrb_state *, const char *)) lookup("mrb_str_new_cstr"))) return -1;
-  return 0;
+    drb_api->mrb_define_module_function(state, module, "ice_cpu_get_info", drb_ffi_ice_cpu_get_info_Binding, MRB_ARGS_REQ(1));
+    struct RClass *ice_cpu_info_ptr_class = drb_api->mrb_define_class_under(state, module, "ICE_CPU_INFO_PTR", object_class);
+    drb_api->mrb_define_class_method(state, ice_cpu_info_ptr_class, "new", drb_ffi__ZTSP12ice_cpu_info_New, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, ice_cpu_info_ptr_class, "value", drb_ffi__ZTSP12ice_cpu_info_GetValue, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, ice_cpu_info_ptr_class, "[]", drb_ffi__ZTSP12ice_cpu_info_GetAt, MRB_ARGS_REQ(1));
+    drb_api->mrb_define_method(state, ice_cpu_info_ptr_class, "[]=", drb_ffi__ZTSP12ice_cpu_info_SetAt, MRB_ARGS_REQ(2));
+    drb_api->mrb_define_method(state, ice_cpu_info_ptr_class, "nil?", drb_ffi__ZTSP12ice_cpu_info_IsNil, MRB_ARGS_REQ(0));
+    struct RClass *CharPointerClass = drb_api->mrb_define_class_under(state, module, "CharPointer", object_class);
+    drb_api->mrb_define_class_method(state, CharPointerClass, "new", drb_ffi__ZTSPc_New, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, CharPointerClass, "value", drb_ffi__ZTSPc_GetValue, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, CharPointerClass, "[]", drb_ffi__ZTSPc_GetAt, MRB_ARGS_REQ(1));
+    drb_api->mrb_define_method(state, CharPointerClass, "[]=", drb_ffi__ZTSPc_SetAt, MRB_ARGS_REQ(2));
+    drb_api->mrb_define_method(state, CharPointerClass, "nil?", drb_ffi__ZTSPc_IsNil, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, CharPointerClass, "str", drb_ffi__ZTSPc_GetString, MRB_ARGS_REQ(0));
+    struct RClass *ice_cpu_info_class = drb_api->mrb_define_class_under(state, module, "ICE_CPU_INFO", object_class);
+    drb_api->mrb_define_class_method(state, ice_cpu_info_class, "new", drb_ffi__ZTS12ice_cpu_info_New, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, ice_cpu_info_class, "name", drb_ffi__ZTS12ice_cpu_info_name_Get, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, ice_cpu_info_class, "name=", drb_ffi__ZTS12ice_cpu_info_name_Set, MRB_ARGS_REQ(1));
+    drb_api->mrb_define_method(state, ice_cpu_info_class, "cores", drb_ffi__ZTS12ice_cpu_info_cores_Get, MRB_ARGS_REQ(0));
+    drb_api->mrb_define_method(state, ice_cpu_info_class, "cores=", drb_ffi__ZTS12ice_cpu_info_cores_Set, MRB_ARGS_REQ(1));
 }